requirements.txt
apache-airflow
psycopg2-binary
SQLAlchemy
requests
pandas
python-dotenv
pyspark
Flask
